# common/api/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/user/avatar/{name}:
        get:
            tags:
                - UserUserService
            description: 获取用户默认头像
            operationId: UserUserService_Avatar
            parameters:
                - name: name
                  in: path
                  description: 用户名称
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.user.v1.AvatarReply'
    /v1/user/getCurrentUser:
        post:
            tags:
                - UserUserService
            description: 获取当前用户信息
            operationId: UserUserService_GetCurrentUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.user.v1.GetCurrentUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.user.v1.GetCurrentUserReply'
    /v1/user/getList:
        post:
            tags:
                - UserUserService
            description: 查询用户列表（返回数组）
            operationId: UserUserService_GetList
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.user.v1.GetListRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.user.v1.GetListReply'
    /v1/user/getMap:
        post:
            tags:
                - UserUserService
            description: 查询用户列表（返回map）
            operationId: UserUserService_GetMap
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.user.v1.GetMapRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.user.v1.GetMapReply'
    /v1/user/getOne:
        post:
            tags:
                - UserUserService
            description: 查询单个用户
            operationId: UserUserService_GetOne
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.user.v1.GetOneRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.user.v1.GetOneReply'
    /v1/user/page:
        post:
            tags:
                - UserUserService
            description: 分页查询用户
            operationId: UserUserService_Page
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.user.v1.PageUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.user.v1.PageUserReply'
    /v1/user/updateSetting:
        post:
            tags:
                - UserUserService
            description: 更新用户设置
            operationId: UserUserService_UpdateSetting
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/common.api.user.v1.UpdateSettingRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/common.api.user.v1.UpdateSettingReply'
components:
    schemas:
...
        common.api.common.v1.PageReply:
            type: object
            properties:
                total:
                    type: integer
                    description: 总数
                    format: uint32
                page:
                    type: integer
                    description: 页码
                    format: uint32
                size:
                    type: integer
                    description: 页大小
                    format: uint32
        common.api.common.v1.PageRequest:
            type: object
            properties:
                page:
                    type: integer
                    description: 页码
                    format: uint32
                size:
                    type: integer
                    description: 页大小
                    format: uint32
...
        common.api.user.v1.AvatarReply:
            type: object
            properties:
                data:
                    type: string
                    description: 图片数据
                    format: bytes
                content_type:
                    type: string
                    description: 图片类型
...
        common.api.user.v1.GetCurrentUserReply:
            type: object
            properties:
                user:
                    allOf:
                        - $ref: '#/components/schemas/common.api.user.v1.User'
                    description: 用户信息
        common.api.user.v1.GetCurrentUserRequest:
            type: object
            properties: {}
        common.api.user.v1.GetListReply:
            type: object
            properties:
                users:
                    type: array
                    items:
                        $ref: '#/components/schemas/common.api.user.v1.User'
                    description: 用户列表
        common.api.user.v1.GetListRequest:
            type: object
            properties:
                query:
                    allOf:
                        - $ref: '#/components/schemas/common.api.user.v1.UserQueryParams'
                    description: 查询参数
        common.api.user.v1.GetMapReply:
            type: object
            properties:
                users:
                    type: object
                    additionalProperties:
                        $ref: '#/components/schemas/common.api.user.v1.User'
                    description: 用户Id -> 用户信息
        common.api.user.v1.GetMapRequest:
            type: object
            properties:
                query:
                    allOf:
                        - $ref: '#/components/schemas/common.api.user.v1.UserQueryParams'
                    description: 查询参数
        common.api.user.v1.GetOneReply:
            type: object
            properties:
                user:
                    allOf:
                        - $ref: '#/components/schemas/common.api.user.v1.User'
                    description: 用户信息
        common.api.user.v1.GetOneRequest:
            type: object
            properties:
                user_id:
                    type: string
                    description: 用户Id
                name:
                    type: string
                    description: 名称
...
        common.api.user.v1.PageUserReply:
            type: object
            properties:
                page:
                    allOf:
                        - $ref: '#/components/schemas/common.api.common.v1.PageReply'
                    description: 分页
                rows:
                    type: array
                    items:
                        $ref: '#/components/schemas/common.api.user.v1.User'
                    description: 用户列表
        common.api.user.v1.PageUserRequest:
            type: object
            properties:
                page:
                    allOf:
                        - $ref: '#/components/schemas/common.api.common.v1.PageRequest'
                    description: 分页参数
                query:
                    allOf:
                        - $ref: '#/components/schemas/common.api.user.v1.UserQueryParams'
                    description: 查询参数
...
        common.api.user.v1.UpdateSettingReply:
            type: object
            properties:
                user:
                    allOf:
                        - $ref: '#/components/schemas/common.api.user.v1.User'
                    description: 用户信息
        common.api.user.v1.UpdateSettingRequest:
            type: object
            properties:
                avatar_url:
                    type: string
                    description: 头像URL
                language:
                    type: string
                    description: 用户语言
                timezone:
                    type: string
                    description: 时区
                theme:
                    type: string
                    description: 皮肤主题
                mobile_theme:
                    type: string
                    description: 移动端皮肤主题
                enable_web_notify:
                    type: boolean
                    description: 启用Web通知
                enable_email_subscribe:
                    type: boolean
                    description: 启用邮件订阅
...
        common.api.user.v1.User:
            type: object
            properties:
                id:
                    type: string
                    description: 用户Id
                name:
                    type: string
                    description: 用户名
                nickname:
                    type: string
                    description: 昵称
                email:
                    type: string
                    description: 邮箱
                phone:
                    type: string
                    description: 手机号
                url:
                    type: string
                    description: 用户个人主页链接
                avatar_url:
                    type: string
                    description: 头像URL
                introduction:
                    type: string
                    description: 个人简介
                mbti:
                    type: string
                    description: 用户 MBTI 类型
                status:
                    type: integer
                    description: 用户状态：0-正常，1-封禁，2-注销
                    format: int32
                group_name:
                    type: string
                    description: 用户组名称
                follow_count:
                    type: integer
                    description: 关注数
                    format: int32
                follower_count:
                    type: integer
                    description: 粉丝数
                    format: int32
                last_login_time:
                    type: string
                    description: 最近登录时间
                    format: date-time
                last_login_ip:
                    type: string
                    description: 最近登录IP
                online_minutes:
                    type: integer
                    description: 在线总时长（分钟）
                    format: int32
                last_checkin_time:
                    type: string
                    description: 最近签到时间
                    format: date-time
                current_checkin_streak:
                    type: integer
                    description: 当前连续签到天数
                    format: int32
                longest_checkin_streak:
                    type: integer
                    description: 最长连续签到天数
                    format: int32
                language:
                    type: string
                    description: 用户语言
                timezone:
                    type: string
                    description: 时区
                theme:
                    type: string
                    description: 皮肤主题
                mobile_theme:
                    type: string
                    description: 移动端皮肤主题
                enable_web_notify:
                    type: boolean
                    description: 启用Web通知
                enable_email_subscribe:
                    type: boolean
                    description: 启用邮件订阅
                public_points:
                    type: boolean
                    description: 是否公开积分榜
                public_followers:
                    type: boolean
                    description: 是否公开粉丝列表
                public_articles:
                    type: boolean
                    description: 是否公开帖子列表
                public_comments:
                    type: boolean
                    description: 是否公开评论列表
                public_online_status:
                    type: boolean
                    description: 是否公开在线状态
                country:
                    type: string
                    description: 所在国家
                province:
                    type: string
                    description: 所在省份
                city:
                    type: string
                    description: 所在城市
                public_location:
                    type: boolean
                    description: 是否公开地理位置
                twofa_secret:
                    type: string
                    description: 二步验证 Secret
                created_at:
                    type: string
                    description: 创建时间
                    format: date-time
                updated_at:
                    type: string
                    description: 更新时间
                    format: date-time
        common.api.user.v1.UserQueryParams:
            type: object
            properties:
                user_id:
                    type: string
                    description: 用户Id
                user_ids:
                    type: array
                    items:
                        type: string
                    description: 用户Id列表
                name:
                    type: string
                    description: 用户名
                names:
                    type: array
                    items:
                        type: string
                    description: 用户名列表
                nickname:
                    type: string
                    description: 昵称
                nicknames:
                    type: array
                    items:
                        type: string
                    description: 昵称列表
                email:
                    type: string
                    description: 邮箱
                emails:
                    type: array
                    items:
                        type: string
                    description: 邮箱列表
                phone:
                    type: string
                    description: 手机号
                phones:
                    type: array
                    items:
                        type: string
                    description: 手机号列表
...