# common/api/user/v1/user_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v6.33.1
// source: user/v1/user.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationUserUserServiceAvatar = "/common.api.user.v1.UserUserService/Avatar"
const OperationUserUserServiceGetCurrentUser = "/common.api.user.v1.UserUserService/GetCurrentUser"
const OperationUserUserServiceGetList = "/common.api.user.v1.UserUserService/GetList"
const OperationUserUserServiceGetMap = "/common.api.user.v1.UserUserService/GetMap"
const OperationUserUserServiceGetOne = "/common.api.user.v1.UserUserService/GetOne"
const OperationUserUserServicePage = "/common.api.user.v1.UserUserService/Page"
const OperationUserUserServiceUpdateSetting = "/common.api.user.v1.UserUserService/UpdateSetting"

type UserUserServiceHTTPServer interface {
	// Avatar 获取用户默认头像
	Avatar(context.Context, *AvatarRequest) (*AvatarReply, error)
	// GetCurrentUser 获取当前用户信息
	GetCurrentUser(context.Context, *GetCurrentUserRequest) (*GetCurrentUserReply, error)
	// GetList 查询用户列表（返回数组）
	GetList(context.Context, *GetListRequest) (*GetListReply, error)
	// GetMap 查询用户列表（返回map）
	GetMap(context.Context, *GetMapRequest) (*GetMapReply, error)
	// GetOne 查询单个用户
	GetOne(context.Context, *GetOneRequest) (*GetOneReply, error)
	// Page 分页查询用户
	Page(context.Context, *PageUserRequest) (*PageUserReply, error)
	// UpdateSetting 更新用户设置
	UpdateSetting(context.Context, *UpdateSettingRequest) (*UpdateSettingReply, error)
}

func RegisterUserUserServiceHTTPServer(s *http.Server, srv UserUserServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/user/updateSetting", _UserUserService_UpdateSetting0_HTTP_Handler(srv))
	r.POST("/v1/user/getCurrentUser", _UserUserService_GetCurrentUser0_HTTP_Handler(srv))
	r.POST("/v1/user/getOne", _UserUserService_GetOne0_HTTP_Handler(srv))
	r.POST("/v1/user/getList", _UserUserService_GetList0_HTTP_Handler(srv))
	r.POST("/v1/user/getMap", _UserUserService_GetMap0_HTTP_Handler(srv))
	r.POST("/v1/user/page", _UserUserService_Page1_HTTP_Handler(srv))
	r.GET("/v1/user/avatar/{name}", _UserUserService_Avatar0_HTTP_Handler(srv))
}

func _UserUserService_UpdateSetting0_HTTP_Handler(srv UserUserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateSettingRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserUserServiceUpdateSetting)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateSetting(ctx, req.(*UpdateSettingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateSettingReply)
		return ctx.Result(200, reply)
	}
}

func _UserUserService_GetCurrentUser0_HTTP_Handler(srv UserUserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetCurrentUserRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserUserServiceGetCurrentUser)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetCurrentUser(ctx, req.(*GetCurrentUserRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetCurrentUserReply)
		return ctx.Result(200, reply)
	}
}

func _UserUserService_GetOne0_HTTP_Handler(srv UserUserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetOneRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserUserServiceGetOne)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetOne(ctx, req.(*GetOneRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetOneReply)
		return ctx.Result(200, reply)
	}
}

func _UserUserService_GetList0_HTTP_Handler(srv UserUserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetListRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserUserServiceGetList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetList(ctx, req.(*GetListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetListReply)
		return ctx.Result(200, reply)
	}
}

func _UserUserService_GetMap0_HTTP_Handler(srv UserUserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetMapRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserUserServiceGetMap)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetMap(ctx, req.(*GetMapRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetMapReply)
		return ctx.Result(200, reply)
	}
}

func _UserUserService_Page1_HTTP_Handler(srv UserUserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in PageUserRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserUserServicePage)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Page(ctx, req.(*PageUserRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*PageUserReply)
		return ctx.Result(200, reply)
	}
}

func _UserUserService_Avatar0_HTTP_Handler(srv UserUserServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in AvatarRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserUserServiceAvatar)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Avatar(ctx, req.(*AvatarRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*AvatarReply)
		return ctx.Result(200, reply)
	}
}

type UserUserServiceHTTPClient interface {
	// Avatar 获取用户默认头像
	Avatar(ctx context.Context, req *AvatarRequest, opts ...http.CallOption) (rsp *AvatarReply, err error)
	// GetCurrentUser 获取当前用户信息
	GetCurrentUser(ctx context.Context, req *GetCurrentUserRequest, opts ...http.CallOption) (rsp *GetCurrentUserReply, err error)
	// GetList 查询用户列表（返回数组）
	GetList(ctx context.Context, req *GetListRequest, opts ...http.CallOption) (rsp *GetListReply, err error)
	// GetMap 查询用户列表（返回map）
	GetMap(ctx context.Context, req *GetMapRequest, opts ...http.CallOption) (rsp *GetMapReply, err error)
	// GetOne 查询单个用户
	GetOne(ctx context.Context, req *GetOneRequest, opts ...http.CallOption) (rsp *GetOneReply, err error)
	// Page 分页查询用户
	Page(ctx context.Context, req *PageUserRequest, opts ...http.CallOption) (rsp *PageUserReply, err error)
	// UpdateSetting 更新用户设置
	UpdateSetting(ctx context.Context, req *UpdateSettingRequest, opts ...http.CallOption) (rsp *UpdateSettingReply, err error)
}

type UserUserServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewUserUserServiceHTTPClient(client *http.Client) UserUserServiceHTTPClient {
	return &UserUserServiceHTTPClientImpl{client}
}

// Avatar 获取用户默认头像
func (c *UserUserServiceHTTPClientImpl) Avatar(ctx context.Context, in *AvatarRequest, opts ...http.CallOption) (*AvatarReply, error) {
	var out AvatarReply
	pattern := "/v1/user/avatar/{name}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserUserServiceAvatar))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetCurrentUser 获取当前用户信息
func (c *UserUserServiceHTTPClientImpl) GetCurrentUser(ctx context.Context, in *GetCurrentUserRequest, opts ...http.CallOption) (*GetCurrentUserReply, error) {
	var out GetCurrentUserReply
	pattern := "/v1/user/getCurrentUser"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserUserServiceGetCurrentUser))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetList 查询用户列表（返回数组）
func (c *UserUserServiceHTTPClientImpl) GetList(ctx context.Context, in *GetListRequest, opts ...http.CallOption) (*GetListReply, error) {
	var out GetListReply
	pattern := "/v1/user/getList"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserUserServiceGetList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetMap 查询用户列表（返回map）
func (c *UserUserServiceHTTPClientImpl) GetMap(ctx context.Context, in *GetMapRequest, opts ...http.CallOption) (*GetMapReply, error) {
	var out GetMapReply
	pattern := "/v1/user/getMap"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserUserServiceGetMap))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetOne 查询单个用户
func (c *UserUserServiceHTTPClientImpl) GetOne(ctx context.Context, in *GetOneRequest, opts ...http.CallOption) (*GetOneReply, error) {
	var out GetOneReply
	pattern := "/v1/user/getOne"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserUserServiceGetOne))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Page 分页查询用户
func (c *UserUserServiceHTTPClientImpl) Page(ctx context.Context, in *PageUserRequest, opts ...http.CallOption) (*PageUserReply, error) {
	var out PageUserReply
	pattern := "/v1/user/page"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserUserServicePage))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateSetting 更新用户设置
func (c *UserUserServiceHTTPClientImpl) UpdateSetting(ctx context.Context, in *UpdateSettingRequest, opts ...http.CallOption) (*UpdateSettingReply, error) {
	var out UpdateSettingReply
	pattern := "/v1/user/updateSetting"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserUserServiceUpdateSetting))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}